all bird ents now are float

+++ Engine/Bird.h
@@ -5,21 +5,21 @@
 class Bird
 {
 public:
-	Bird(int _x, int _y);
+	Bird(float _x, float _y);
 	void DrawBird(Graphics& gfx);
 	void Move();
 	void ClampToScreen();
-	int BirdXSize();
-	int BirdYSize();
-	int GetBirdX();
+	float BirdXSize();
+	float BirdYSize();
+	float GetBirdX();
 	float GetBirdY();
 	bool isJumped();
 private:
-	int x;
+	float x;
 	float y;
-	static constexpr int width = 70;
-	static constexpr int height = 53;
-	float vy = 1;
+	static constexpr float width = 70.0f;
+	static constexpr float height = 53.0f;
+	float vy = 1.0f;
 	bool enable = false;
 	bool inhibit = false;
 	bool soundPlay = false;
